Added a RandomStrategy and corresponding traits, structs and adjustments

+++ trader/src/strategies/decision.rs
@@ -1,10 +1,12 @@
 use rust_decimal::Decimal;
 
+#[derive(Debug)]
 pub struct Position {
     pub price: Decimal,
-    pub volume: f32,
+    pub volume: Decimal,
 }
 
+#[derive(Debug)]
 pub enum Decision {
     Buying { position: Position },
     Holding,
